Fix test case

+++ j--/tests/pass/For.java
@@ -25,7 +25,7 @@ public class For {
     }
     public static int doubleFor(int a){
         int c = 0;
-        for(double b = 0.0; b > a; b += 1.0 ){
+        for(double b = 0.0; b > (double)a; b += 1.0 ){
             c += 1;
         }
         return c;
